[msvc] compile fixes

+++ src/Libpfs/vex/dotproduct.hxx
@@ -34,7 +34,7 @@ _Type dotProduct(const _Type* v1, const _Type* v2, size_t N)
 {
     _Type dotProd = _Type();
 #pragma omp parallel for reduction(+:dotProd)
-    for (size_t idx = 0; idx < N; idx++)
+    for (int idx = 0; idx < N; idx++)
     {
         dotProd += (v1[idx] * v2[idx]);
     }
@@ -46,7 +46,7 @@ _Type dotProduct(const _Type* v1, size_t N)
 {
     _Type dotProd = _Type();
 #pragma omp parallel for reduction(+:dotProd)
-    for (size_t idx = 0; idx < N; idx++)
+    for (int idx = 0; idx < N; idx++)
     {
         dotProd += (v1[idx] * v1[idx]);
     }
